requirements.txt
gradio
pillow
fastai
timm